[PATCH] early uart console support
This adds an early polled-mode "uart" console driver, based on Andi Kleen's
early_printk work.

The difference is that this locates the UART device directly by its MMIO or
I/O port address, so we don't have to make assumptions about how ttyS
devices will be named.  After the normal serial driver starts, we try to
locate the matching ttyS device and start a console there.

Sample usage:
    console=uart,io,0x3f8
    console=uart,mmio,0xff5e0000,115200n8

If the baud rate isn't specified, we peek at the UART to figure it out.

Signed-off-by: Bjorn Helgaas <bjorn.helgaas@hp.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>