[PATCH] fix for breakage introduced in hgafb.c
ChangeSet "adaplas: [PATCH] fbdev: Add __iomem annotations to hgafb" had
actually broken it - it not only added a (useless) cast in rowaddr(), it
had also added a dereferencing.

Since it is video memory and since we do memory transfers on the
resulting "address"...  IOW, it's even exploitable - not that there had
been a lot of HGA boxen out there, but... 

This restores the intended behaviour.

Signed-off-by: Al Viro <viro@parcelfarce.linux.theplanet.co.uk>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>