[EBTABLES]: Add userspace logging via netlink socket.
The patch below adds the ebtables ulog watcher, that sends packets
to userspace. It is based on ipt_ULOG.c.

The complete packet, including Ethernet header, is sent to userspace.
The in and out bridge ports are also sent to userspace. Logging is
of course not restricted to IP packets.
An example of a userspace program that receives and parses packets
sent by the ulog watcher is in the ebtables CVS tree under
examples/ulog/

Signed-off-by: Bart De Schuymer <bdschuym@pandora.be>
Signed-off-by: David S. Miller <davem@davemloft.net>