[PATCH] USB: add ati_remote.c device id
please consider applying the attached patch which adds an identifier for
another variant of the X10 RF remote. This one is very common everywhere
in Europe where Aldi (a supermarket discounter) sells their PCs
(Netherlands, Belgium, northern Germany, France, Spain, ...).
Considering that they ship a few 100,000 PCs once or twice every year
you can imagine there are quite a few of these remotes kicking around.

I've tested with 2.6.10-rc2 and various versions of XMMS and lirc with
the devinput driver. It does play well with lirc-0.7.0-pre8 but not so
nice with 0.7.0. I don't know why, but according to some posts on the
lirc list it seems to be in lirc.

By the way, there seem to be even more variants of this device,
according to:

http://cvs.sourceforge.net/viewcvs.py/lirc/lirc/drivers/lirc_atiusb/lirc_atiusb.c

From: Armijn Hemel <armijn@uulug.nl>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>