[PATCH] ppc64: fix compilation with recent toolchains
The ppc64 toolchains don't create dot symbols (i.e. a globally
visible ".foo" symbol for the text of function foo) any more.

This breaks the kernel compile because we refer to function text
addresses in the system call table.

Fortunately there is an option, -mcall-aixdesc, which restores the
previous behaviour, and even more fortunately, old ppc64 toolchains
understand the option as well as new ones.

This patch adds -mcall-aixdesc to CFLAGS in arch/ppc64/Makefile.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>