[PATCH] ppc32: Fix an IRQ issue with cpufreq
The ppc32 PowerMac cpufreq code, when using the PMU to switch the
frequency, would eventually lose interrupts.  The solution is to raise the
CPU priority at the controller level.  It's also unnecessary to call the
full PIC suspend/resume code in this case as the IO chip isn't reset,
unlike the sleep code.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>