[ALSA] fix display of send routing in /proc
EMU10K1/EMU10K2 driver
The emu10k1's send routing is a per channel setting.  The emu10k1 has 4
sends per channel, the Audigy 8.  Currently we only display the first 4
sends for the first channel.  This patch updates the /proc file to
display the send routing for all 64 channels, and to display the last 4
sends if present.

Signed-off-by: Lee Revell <rlrevell@joe-job.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>