[ARM PATCH] 2290/1: S3C2410 - timer usec fixes
Patch from Ben Dooks

A simple divide is not enough for the s3c2410 default
timer code, due to most clocks not being a nice multiple
of usecs.

This update changes the calculation of the usec to use
multipliers and shifts to effect a fast divide by a non-integer
number.

Original patch by Dimitry Andric, updated by Ben Dooks

Signed-off-by: Dimitry Andric

Signed-off-by: Ben Dooks
Signed-off-by: Russell King