[ACPI] ACPICA 20041203 from Bob Moore and Alexey Starikovskiy
The low-level field insertion/extraction code (exfldio)
has been completely rewritten to eliminate unnecessary
complexity, bugs, and boundary conditions.

Fixed a problem in the ToInteger, ToBuffer, ToHexString,
and ToDecimalString operators where the input operand could
be inadvertently deleted if no conversion was necessary
(e.g., if the input to ToInteger was an Integer object.)

Fixed a problem with the ToDecimalString and ToHexString
where an incorrect exception code was returned if the
resulting string would be > 200 chars.  AE_STRING_LIMIT is
now returned.

Fixed a problem with the Concatenate operator where AE_OK
was always returned, even if the operation failed.

Fixed a problem in oswinxf (used by AcpiExec and iASL)
to allow > 128 semaphores to be allocated.

Signed-off-by: Len Brown <len.brown@intel.com>