[PATCH] hack imm.c to work in highmem machines
Currently imm uses page_address() which can crash on highmem. It's not
directly doable to map the pages properly, at least not without
changing some code. In lack of a ->bounce_highio member in the scsi
host template, just set ->unchecked_isa_dma which will just bounce
everything for us. imm isn't performance critical by any stretch of
the imagination, so...

Usually I'd not encourage such a silly hack, but in lack of hardware for
testing (who has it??), this should suffice as it is obviously correct.

Signed-off-by: Jens Axboe <axboe@suse.de>
Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>