[PATCH] ppc64: make sure lppaca doesn't cross page boundary
This patch is from Will Schmidt <willschm@us.ibm.com>, with some extra
comments from me.

On iSeries and on POWER5 machines, there is a data structure which is
used for communication between the hypervisor and the kernel, called
the `lppaca'.  The kernel tells the hypervisor where it is, and the
hypervisor requires that it doesn't cross a page boundary.  With other
changes in the last few months we have ended up with a situation where
it could cross a page boundary.  This patch increases the alignment
requirement for the struct to make sure that it can't cross a page
boundary.

This is a bug fix and should go into 2.6.10.

Signed-off-by:  Will Schmidt  <willschm@us.ibm.com>
Signed-off-by:  Paul Mackerras <paulus@samba.org>
Signed-off-by:  Linus Torvalds <torvalds@osdl.org>