[PATCH] ppc64: make UP kernel run on POWER4 logical partition
Recently I tried booting a UP kernel on a 1-processor logical partition on
a POWER4 system.  It failed because the CPU we happened to be running on
was CPU 2, but hard_smp_processor_id() is defined to 0 for !CONFIG_SMP. 
(Note that this code runs quite early, as part of init_IRQ, so
hard_smp_processor_id() should be the physical ID of the boot cpu.)

This patch does a minimal fix to make it work.  I think this patch should
go in 2.6.10.  Ultimately I think the hard_smp_processor_id() definition
should be removed from include/linux/smp.h, but that carries a risk of
breaking other architectures and probably shouldn't be done pre 2.6.10.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>