[PATCH] USB: Hub driver cleanups [12/13]
This patch does a little cleanup of the hub driver, nothing dramatic
(although it is long because it touches a lot of code).  The main feature
is that the private hub structure is passed as an argument to internal
routines rather than the hub device structure.  This simplifies
conversions, because it's easier to go from the private structure to the
device structure than vice versa.  The other change is to store a pointer
to the interface's embedded struct device rather than the interface
itself.  This helps because that pointer is mainly used for dev_xxx log
messages.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Greg Kroah-Hartman <greg@kroah.com>